+++ geotrek/common/tests/test_conf.py
@@ -2,7 +2,8 @@ from django.conf import settings
 from django.core.checks import Error
 from django.test import TestCase
 
-from geotrek.common.apps import CommonConfig, srid_check
+from geotrek.common.apps import CommonConfig
+from geotrek.common.checks import srid_check
 
 
 class StartupCheckTest(TestCase):