+++ geotrek/maintenance/urls.py
@@ -1,10 +1,10 @@
-from django.conf import settings
 from django.urls import path
 from mapentity.registry import registry
 
-from . import models, views
+from . import entities, models, views
 
 app_name = "maintenance"
+
 urlpatterns = [
     path(
         "api/intervention/references/",
@@ -14,6 +14,6 @@ urlpatterns = [
 ]
 
 urlpatterns += registry.register(
-    models.Intervention, menu=settings.INTERVENTION_MODEL_ENABLED
+    models.Intervention, options=entities.InterventionEntityOptions
 )
-urlpatterns += registry.register(models.Project, menu=settings.PROJECT_MODEL_ENABLED)
+urlpatterns += registry.register(models.Project, options=entities.ProjectEntityOptions)