+++ geotrek/signage/views.py
@@ -100,6 +100,7 @@ class SignageFormatList(MapEntityFormat, SignageList):
         "access",
         "manager",
         "uuid",
+        "coupled",
         *AltimetryMixin.COLUMNS,
     ]