+++ geotrek/trekking/tests/test_views.py
@@ -2,7 +2,7 @@ import csv
 import os
 from collections import OrderedDict
 from io import StringIO
-from unittest import mock, skipIf
+from unittest import mock
 
 from bs4 import BeautifulSoup
 from django.conf import settings
@@ -97,11 +97,9 @@ class POIViewsTest(CommonTest):
             "description_en": "here",
             "type": POITypeFactory.create().pk,
         }
-        if settings.TREKKING_TOPOLOGY_ENABLED:
-            PathFactory.create()
-            good_data["topology"] = '{"lat": 5.1, "lng": 6.6}'
-        else:
-            good_data["geom"] = "POINT(5.1 6.6)"
+        PathFactory.create()
+        good_data["topology"] = '{"type":"Point","coordinates":[6.6,5.1]}'
+
         return good_data
 
     def get_expected_popup_content(self):
@@ -111,6 +109,9 @@ class POIViewsTest(CommonTest):
             f'    <a id="detail-btn" href="/poi/{self.obj.pk}/" class="btn btn-sm btn-info mt-2">Detail sheet</a>\n'
             f"</div>"
         )
+
+    def _check_update_geom_permission(self, response):
+        pass
 
     def test_status_only_review(self):
         element_not_published = self.modelfactory.create(published=False, review=True)
@@ -155,17 +156,12 @@ class POIViewsTest(CommonTest):
 
     def test_empty_topology(self):
         data = self.get_good_data()
-        if settings.TREKKING_TOPOLOGY_ENABLED:
-            data["topology"] = ""
-        else:
-            data["geom"] = ""
+        data["topology"] = ""
+
         response = self.client.post(self.model.get_add_url(), data)
         self.assertEqual(response.status_code, 200)
         form = self.get_form(response)
-        if settings.TREKKING_TOPOLOGY_ENABLED:
-            self.assertEqual(form.errors, {"topology": ["Topology is empty."]})
-        else:
-            self.assertEqual(form.errors, {"geom": ["No geometry value provided."]})
+        self.assertEqual(form.errors, {"topology": ["Topology is empty."]})
 
     def test_listing_number_queries(self):
         # Create many instances
@@ -318,14 +314,10 @@ class TrekViewsTest(CommonTest):
             "web_links": WebLinkFactory.create().pk,
             "information_desks": tourism_factories.InformationDeskFactory.create().pk,
         }
-        if settings.TREKKING_TOPOLOGY_ENABLED:
-            good_data["topology"] = f'{{"paths": [{self.path.pk}]}}'
-            good_data["pois_excluded"] = POIFactory.create(paths=[self.path]).pk
-        else:
-            good_data["geom"] = "SRID=4326;LINESTRING (0.0 0.0, 1.0 1.0)"
-            good_data["pois_excluded"] = POIFactory.create(
-                geom="SRID=2154;POINT (700000 6600000)"
-            ).pk
+        good_data["topology"] = f'{{"paths": [{self.path.pk}]}}'
+        good_data["topology_changed"] = "true"
+        good_data["pois_excluded"] = POIFactory.create(paths=[self.path]).pk
+
         return good_data
 
     def get_expected_popup_content(self):
@@ -345,7 +337,7 @@ class TrekViewsTest(CommonTest):
 
     def test_badfield_goodgeom(self):
         bad_data, form_error = self.get_bad_data()
-        bad_data["parking_location"] = "POINT (1.0 1.0)"  # good data
+        bad_data["parking_location"] = "SRID=4326;POINT (1.0 1.0)"  # good data
 
         url = self.model.get_add_url()
         response = self.client.post(url, bad_data)
@@ -368,21 +360,18 @@ class TrekViewsTest(CommonTest):
         self.assertNotContains(response, "pois_excluded")
 
     def test_pois_detached_update(self):
-        if settings.TREKKING_TOPOLOGY_ENABLED:
-            p1 = PathFactory.create(geom=LineString((0, 0), (4, 4)))
-            trek = TrekFactory.create(paths=[p1])
-            poi = POIFactory.create(paths=[(p1, 0.6, 0.6)])
-        else:
-            trek = TrekFactory.create(geom="SRID=4326;LINESTRING (0.0 0.0, 1.0 1.0)")
-            poi = POIFactory.create(geom="SRID=4326;POINT (0.6 0.6)")
+        p1 = PathFactory.create(geom=LineString((0, 0), (4, 4)))
+        trek = TrekFactory.create(paths=[p1])
+        poi = POIFactory.create(paths=[(p1, 0.6, 0.6)])
+
         good_data = self.get_good_data()
         good_data["pois_excluded"] = poi.pk
         self.client.post(self.model.get_update_url(trek), good_data)
         self.assertIn(poi, trek.pois_excluded.all())
 
-    def test_detail_lother_language(self):
+    def test_detail_other_language(self):
         bad_data, form_error = self.get_bad_data()
-        bad_data["parking_location"] = "POINT (1.0 1.0)"  # good data
+        bad_data["parking_location"] = "SRID=4326;POINT (1.0 1.0)"  # good data
 
         url = self.model.get_add_url()
         response = self.client.post(url, bad_data)
@@ -406,15 +395,8 @@ class TrekViewsTest(CommonTest):
             ),
             "parking_location": Point(0, 0, srid=settings.SRID),
         }
-        if settings.TREKKING_TOPOLOGY_ENABLED:
-            path1 = PathFactory.create(
-                geom=f"SRID={settings.SRID};LINESTRING(0 0, 1 0)"
-            )
-            self.trek = TrekFactory.create(paths=[path1], **trek_args)
-        else:
-            self.trek = TrekFactory.create(
-                geom=f"SRID={settings.SRID};LINESTRING(0 0, 1 0)", **trek_args
-            )
+        path1 = PathFactory.create(geom=f"SRID={settings.SRID};LINESTRING(0 0, 1 0)")
+        self.trek = TrekFactory.create(paths=[path1], **trek_args)
         fmt = "csv"
         response = self.client.get(self.model.get_format_list_url() + "?format=" + fmt)
         self.assertEqual(response.status_code, 200)
@@ -982,17 +964,11 @@ class ServiceViewsTest(CommonTest):
         }
 
     def get_good_data(self):
-        if settings.TREKKING_TOPOLOGY_ENABLED:
-            PathFactory.create()
-            return {
-                "type": ServiceTypeFactory.create().pk,
-                "topology": '{"lat": 5.1, "lng": 6.6}',
-            }
-        else:
-            return {
-                "type": ServiceTypeFactory.create().pk,
-                "geom": "POINT(5.1 6.6)",
-            }
+        PathFactory.create()
+        return {
+            "type": ServiceTypeFactory.create().pk,
+            "topology": '{"type":"Point","coordinates":[6.6,5.1]}',
+        }
 
     def get_expected_popup_content(self):
         return (
@@ -1002,9 +978,9 @@ class ServiceViewsTest(CommonTest):
             f"</div>"
         )
 
-    @skipIf(
-        not settings.TREKKING_TOPOLOGY_ENABLED, "Test with dynamic segmentation only"
-    )
+    def _check_update_geom_permission(self, response):
+        pass
+
     def test_empty_topology(self):
         data = self.get_good_data()
         data["topology"] = ""
@@ -1012,17 +988,6 @@ class ServiceViewsTest(CommonTest):
         self.assertEqual(response.status_code, 200)
         form = self.get_form(response)
         self.assertEqual(form.errors, {"topology": ["Topology is empty."]})
-
-    @skipIf(
-        settings.TREKKING_TOPOLOGY_ENABLED, "Test without dynamic segmentation only"
-    )
-    def test_empty_topology_nds(self):
-        data = self.get_good_data()
-        data["geom"] = ""
-        response = self.client.post(self.model.get_add_url(), data)
-        self.assertEqual(response.status_code, 200)
-        form = self.get_form(response)
-        self.assertEqual(form.errors, {"geom": ["No geometry value provided."]})
 
     def test_listing_number_queries(self):
         # Create many instances