+++ docs/changelog.rst
@@ -2,8 +2,31 @@
 CHANGELOG
 =========
 
-2.125.2+dev     (XXXX-XX-XX)
+3.0.0-rc1       (XXXX-XX-XX)
 ----------------------------
+
+**Breaking changes**
+
+* Bye Leaflet, Hello MaplibreGL JS
+* Take a look to https://geotrek.readthedocs.io/en/stable/installation-and-configuration/upgrade.html#from-geotrek-admin-2-113-0-to-3-x
+* The ``SIGNAGE_LINE_ENABLED`` setting, which allows to create linear signage, has been removed.
+
+**Warnings**
+
+* Topologies no longer need to depend on the path network: they can now exist independently or be based on paths, regardless of the dynamic segmentation setting.
+* From now, if a module is disabled by its setting, it is completely hidden from menu, layer tree, admin and API responses.
+* All geotrek modules are now installed by default. Please remove `geotrek.outdoor`, `geotrek.sensitivity`, `geotrek.diving` from you custom.py file. User setting to disable module if you don't want to use them.
+  - `COURSE_MODEL_ENABLED = FALSE`
+  - `SITE_MODEL_ENABLED = FALSE`  # for outdoor module
+  - `SENSITIVE_AREA_MODEL_ENABLED = FALSE`  # for sensitivity module
+  - `DIVE_MODEL_ENABLED = FALSE`  # for diving module
+
+**Improvements**
+
+* Add a "Network-coupled" boolean field and its corresponding filter to the ``Topology`` model
+* Don't break the geometry of a topology when it's decoupled from the path network
+* Strengthen the "invalid topology" filter criteria
+* Remove the "invalid geometry" filter
 
 
 2.125.2         (2026-07-21)
@@ -45,7 +68,6 @@ GTAM
 **Bug fixes**
 
 * Revert filters behavior as before select2 migration
-
 
 
 2.124.6         (2026-06-29)
@@ -111,6 +133,7 @@ GTAM
 **Improvements**
 
 * Remove ``DatabaseBackend`` from core and add ``django-flask-authent-backend`` as external dependencies (#5382)
+* From now, if a module is disabled by its setting, it is completely hidden from menu, layer tree, admin and API responses.
 * Add ``--ids`` option to the ``reorder_topologies`` command to process specific topologies by ID
 * Remove blade double link (keep signage ForeignKey and remove topology ForeignKey) (#3009)